Move `WolfContext*` related functionality into `context.rs`
